# lib/api-zod/src/generated/types/plannerInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlannerInput {
  /** Number of study hours available today */
  study_hours: number;
  /** Current mood of the student */
  mood: string;
  /** Description of recent diet or food intake */
  diet: string;
}