# lib/api-zod/src/generated/types/plannerResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlannerResult {
  /** A personalised study plan based on available hours and mood */
  study_plan: string;
  /** A diet suggestion to support study performance */
  diet_suggestion: string;
  /** An uplifting motivational message */
  motivation_message: string;
}